Reorganize localizations tests
