+++ tests/Localization/Da.php
@@ -1,0 +1,79 @@
+<?php
+
+namespace Tests\Localization;
+
+/*
+ * This file is part of the Carbon package.
+ *
+ * (c) Brian Nesbitt <brian@nesbot.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+use Carbon\Carbon;
+use Tests\AbstractTestCase;
+
+class Da extends AbstractTestCase
+{
+    public function testDiffForHumansLocalizedInDanish()
+    {
+        Carbon::setLocale('da');
+
+        $scope = $this;
+        $this->wrapWithTestNow(function () use ($scope) {
+            $d = Carbon::now()->subSecond();
+            $scope->assertSame('1 sekund siden', $d->diffForHumans());
+
+            $d = Carbon::now()->subSeconds(2);
+            $scope->assertSame('2 sekunder siden', $d->diffForHumans());
+
+            $d = Carbon::now()->subMinute();
+            $scope->assertSame('1 minut siden', $d->diffForHumans());
+
+            $d = Carbon::now()->subMinutes(2);
+            $scope->assertSame('2 minutter siden', $d->diffForHumans());
+
+            $d = Carbon::now()->subHour();
+            $scope->assertSame('1 time siden', $d->diffForHumans());
+
+            $d = Carbon::now()->subHours(2);
+            $scope->assertSame('2 timer siden', $d->diffForHumans());
+
+            $d = Carbon::now()->subDay();
+            $scope->assertSame('1 dag siden', $d->diffForHumans());
+
+            $d = Carbon::now()->subDays(2);
+            $scope->assertSame('2 dage siden', $d->diffForHumans());
+
+            $d = Carbon::now()->subWeek();
+            $scope->assertSame('1 uge siden', $d->diffForHumans());
+
+            $d = Carbon::now()->subWeeks(2);
+            $scope->assertSame('2 uger siden', $d->diffForHumans());
+
+            $d = Carbon::now()->subMonth();
+            $scope->assertSame('1 måned siden', $d->diffForHumans());
+
+            $d = Carbon::now()->subMonths(2);
+            $scope->assertSame('2 måneder siden', $d->diffForHumans());
+
+            $d = Carbon::now()->subYear();
+            $scope->assertSame('1 år siden', $d->diffForHumans());
+
+            $d = Carbon::now()->subYears(2);
+            $scope->assertSame('2 år siden', $d->diffForHumans());
+
+            $d = Carbon::now()->addSecond();
+            $scope->assertSame('om 1 sekund', $d->diffForHumans());
+
+            $d = Carbon::now()->addSecond();
+            $d2 = Carbon::now();
+            $scope->assertSame('1 sekund efter', $d->diffForHumans($d2));
+            $scope->assertSame('1 sekund før', $d2->diffForHumans($d));
+
+            $scope->assertSame('1 sekund', $d->diffForHumans($d2, true));
+            $scope->assertSame('2 sekunder', $d2->diffForHumans($d->addSecond(), true));
+        });
+    }
+}
